Code review batch: 8 engine/load-path fixes (CR-2,3,4,5,6,7,9,10,11,13)
CR-2  GLIDE: flush stale glide bookkeeping on the emitter-disable edge (targeted
      per disabled bus, so glides on still-live emitters survive) + the preview
      activation edge (parity flushMod/flushGlide) - else a reused voice slot
      later emits a spurious note-off + the glide goes silent.
CR-3  ReelDeck: a note-ON with no matching OFF (held across the pass boundary, or
      truncated at histCap when re-selected from the browser) now gets a synthetic
      OFF at the loop end (fixed scratch, no render alloc) - no stuck notes on
      re-selection. +1 ReelDeckTest.
CR-4  Master MUTE now silences the BYPASS monitor (a parallel emission path), not
      just the grid - the whole-instrument mute no longer leaks a live wire.
      +1 RouterTest.
CR-5  Flood-governor capacity check moved BEFORE the MONO steal in emitOneBus - a
      MONO note that the governor then drops no longer closes the existing holder
      (silent emitter for the beat).
CR-6  BURST HITS 12/16 rendered as 8 - the shared "count" snapshot clamp was 2..8;
      widened to 2..16 (RATCHET re-snaps to [2,3,4,6,8] via effectiveRepeats, so
      only BURST unlocks). +1 AcceptanceBurstTest.
CR-7  Resync the parameter tree on the factory/preset/fullState load paths (else
      host automation snaps a fresh note back to the OLD morph/transpose/macros).
CR-9  rackConfigsResolved PADS short / TRUNCATES long arrays to 4 instead of
      discarding them wholesale (config data-loss). +1 SnapshotBuilderTest.
CR-10 fullState.set resets pendingBuildUnassigned so the loaded doc's own
      unassigned part isn't re-encoded over on the next save (data-loss).
CR-11 Multi-clock column trackers (prevEffColumnRow[]) re-init when the render
      LEAVES the uniform fast-path, not only on a flush - fixes a stale-edge
      desync on a live per-part clock-mode switch.
CR-13 SnapCell/Voice colourIndex widened Int8 -> Int16 (a document colour index
      can exceed 127 -> Int8 trapped); feed writes clamped; colour param
      observer/provider/resync guarded to document.colours.count.

macOS 845 green (incl. fuzz); iOS builds.

+++ AUExtension/Snapshot.swift
@@ -26,7 +26,7 @@ enum Snap {
 // MARK: - Flat cell (one per grid position; colourIndex < 0 = empty)
 
 struct SnapCell {
-    var colourIndex: Int8 = -1
+    var colourIndex: Int16 = -1   // CR-13a: Int16 (was Int8) — with the 16-colour cap gone, a document colour index can exceed 127; Int8(index) trapped at ≥128
     var alt = false
     var bypassed = false
     var passthrough = false     // NO-MACHINE chain (Paul 2026-08-23): a LIVE WIRE — its input passes straight through in
